# docs/技术架构/openapi/tradehub.openapi.yaml
openapi: 3.1.0
info:
  title: TradeHub API
  version: 1.0.0
  description: |
    TradeHub 第一阶段 API 契约骨架。旧 `/api/v1/*` 接口继续兼容，
    新增能力按 `/api/{domain}/v1/*` 分域推进。
servers:
  - url: http://127.0.0.1:7002
    description: stock-api local
paths:
...
  /api/stock/v1/watchlist/groups:
    get:
      tags:
        - stock
      summary: 获取股票自选分组
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 创建或更新股票自选分组
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags:
        - stock
      summary: 删除股票自选分组
      parameters:
        - name: group_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/stock/v1/watchlist/items:
    get:
      tags:
        - stock
      summary: 获取股票自选项
      parameters:
        - name: group_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 添加或更新股票自选项
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags:
        - stock
      summary: 删除股票自选项
      parameters:
        - name: item_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
  schemas:
    ApiMeta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
          example: req_20260614_120000_123456
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          example: v1
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: string
        detail:
          type: string
    ApiResponse:
      type: object
      required:
        - success
        - code
        - message
        - data
        - meta
      properties:
        success:
          type: boolean
        code:
          type: string
          example: OK
        message:
          type: string
          example: ok
        data: {}
        meta:
          $ref: "#/components/schemas/ApiMeta"
        error:
          $ref: "#/components/schemas/ApiErrorDetail"
...